+++ frontend/src/components/previewers/audio/manifest.json
@@ -1,0 +1,23 @@
+{
+	"name": "basic-audio-component",
+	"version": "1.0.0",
+	"description": "A React component for audio",
+	"main": "Audio.tsx",
+	"dependencies": {
+		"clowder2-core": "1.0.0",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2"
+	},
+	"vizConfig": {
+		"name": "Audio",
+		"mainType": "audio",
+		"mimeType": [
+			"audio/mpeg",
+			"audio/wav",
+			"audio/*"
+		],
+		"props": {
+			"fileId": "string"
+		}
+	}
+}
