+++ frontend/src/components/previewers/html/manifest.json
@@ -1,0 +1,21 @@
+{
+	"name": "basic-iframe-component",
+	"version": "1.0.0",
+	"description": "A React component for html",
+	"main": "Iframe.tsx",
+	"dependencies": {
+		"clowder2-core": "1.0.0",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2"
+	},
+	"vizConfig": {
+		"name": "Iframe",
+		"mainType": "text",
+		"mimeType": [
+			"text/html"
+		],
+		"props": {
+			"fileId": "string"
+		}
+	}
+}
