+++ frontend/src/components/previewers/table/manifest.json
@@ -1,0 +1,21 @@
+{
+	"name": "basic-table-component",
+	"version": "1.0.0",
+	"description": "A React component for table",
+	"main": "Table.tsx",
+	"dependencies": {
+		"clowder2-core": "1.0.0",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2"
+	},
+	"vizConfig": {
+		"name": "Table",
+		"mainType": "text",
+		"mimeType": [
+			"text/csv"
+		],
+		"props": {
+			"fileId": "string"
+		}
+	}
+}
