+++ frontend/src/components/previewers/video/manifest.json
@@ -1,0 +1,21 @@
+{
+	"name": "basic-video-component",
+	"version": "1.0.0",
+	"description": "A React component for video",
+	"main": "Video.tsx",
+	"dependencies": {
+		"clowder2-core": "1.0.0",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2"
+	},
+	"vizConfig": {
+		"name": "Video",
+		"mainType": "video",
+		"mimeType": [
+			"video/*"
+		],
+		"props": {
+			"fileId": "string"
+		}
+	}
+}
